fix expand button disappear after clicking send
